Full Download! (I think?)
I think this is working, but I'm not sure. There seems to be a lot of CDN links, rather than app bundles. That's unfortunate, but also actually kind of helpful, come to think of it. This makes it really easy to move to npm packages possibly, if those are on npm as well.